Some hacky refactoring, some commenting
- export.py has been hackily refactored and splintered into new files - export, exception, export_thread, and paths.

- emoji.py has been fully documented.
- paths.py has been fully documented.
- manifest.py has been commented some more
- shuffled content in the readme.

Renamed stuff:
exception.SkipException -> exception.FilterException
