+++ README.md
@@ -8,6 +8,10 @@ Introduction
 
 Orxporter is a tool for batch export operations for SVG-based emoji sets, providing multi-format and multi-resolution batch rendering, automated recolouring tools, unicode and shortcode filename export, JSON metadata export and more.
 
+
+Disclaimer
+----------
+
 Orxporter was created as a rewrite of scripts used to automate some of the
 workflow for Mutant Standard. Its main purpose remains to be the primary build
 tool for Mutant Standard packages, and so its features are mainly focused on a
@@ -21,12 +25,6 @@ We welcome suggestions, feature requests and collaboration in order to make
 orxporter more useful for independent emoji creators and end-users. We follow
 Mutant Standard's code of conduct - please refer to **CODE_OF_CONDUCT.md**.
 
-Disclaimer
-----------
-
-Orxporter is an open-sourced insider tool, and wasn't originally meant to be
-used at large. There has been no significant effort to design, test, or follow
-any good software development practice.
 
 Features
 --------
